Migrate out tests, remove old fun calls
