+++ src/wikiman/wikiman.py
@@ -61,20 +61,20 @@ def get_tree(page: Path) -> str:
         page_idx = 0
     else:
         # Tree is a list of the page and its siblings
-        siblings = get_siblings(page)
+        siblings = family.get_siblings(page)
         tree = [utils.get_page_link(page) for page in siblings]
         page_idx = siblings.index(page)
         tree[page_idx] = utils.bold_md(tree[page_idx])
 
     # Insert child links below the page
-    children = get_children(page)
+    children = family.get_children(page)
     child_links = [utils.get_page_link(page) for page in children]
     tree = insert_subtree(subtree=child_links, tree=tree, index=page_idx)
 
     # Only worry about parents if it's not the root page
     if page != common.ROOT_PAGE:
 
-        parent = get_parent(page)
+        parent = family.get_parent(page)
 
         if parent == common.ROOT_PAGE:
             # Insert the working tree below the root page
@@ -82,7 +82,7 @@ def get_tree(page: Path) -> str:
             parent_idx = 0
         else:
             # Insert the working tree below the parent page in the list of parents
-            parents = get_siblings(parent)
+            parents = family.get_siblings(parent)
             parent_links = [utils.get_page_link(page) for page in parents]
             parent_idx = parents.index(parent)
         tree = insert_subtree(subtree=tree, tree=parent_links, index=parent_idx)
@@ -157,54 +157,14 @@ def get_relative_nav(page: Path) -> str:
 def get_nearest(page: Path) -> tuple[Path, Path, Path]:
     """Get the pages nearest to a page."""
 
-    siblings = get_siblings(page)
+    siblings = family.get_siblings(page)
     if page == common.ROOT_PAGE:
         next_page = siblings[0] if siblings else common.ROOT_PAGE
         prev_page = common.ROOT_PAGE
     else:
         page_position = siblings.index(page)
-        next_page = get_next(page, siblings, page_position)
-        prev_page = get_prev(page, siblings, page_position)
+        next_page = utils.get_next(page, siblings, page_position)
+        prev_page = utils.get_prev(page, siblings, page_position)
 
-    parent = get_parent(page)
+    parent = family.get_parent(page)
     return next_page, prev_page, parent
-
-
-def get_next(page: Path, siblings: list[Path], page_position: int) -> Path:
-    """Get the page just after this page."""
-
-    return utils.get_next(page, siblings, page_position)
-
-
-def get_next_of_last_child(page: Path) -> Path:
-    """Get the next page of a last child."""
-
-    return utils.get_next_of_last_child(page)
-
-
-def get_prev(page: Path, siblings: list[Path], page_position: int) -> Path:
-    """Get the page just before this page."""
-
-    return utils.get_prev(page, siblings, page_position)
-
-
-# * -------------------------------------------------------------------------------- * #
-# * RELATIVES
-
-
-def get_siblings(page: Path) -> list[Path]:
-    """Get a page and its siblings. The home page has its children as its siblings."""
-
-    return family.get_siblings(page)
-
-
-def get_parent(page: Path) -> Path:
-    """Get the parent of a page."""
-
-    return family.get_parent(page)
-
-
-def get_children(page: Path) -> list[Path]:
-    """Get the children of a page."""
-
-    return family.get_children(page)
