+++ tests/test_wikiman.py
@@ -129,224 +129,6 @@ PAGES = {str(key).lower(): value / (key + ".md") for key, value in PAGE_PATHS.it
 # ! -------------------------------------------------------------------------------- ! #
 # ! API
 
-# * -------------------------------------------------------------------------------- * #
-# * RELATIVES
-
-# * ---------------------------------------- * #
-# * get_children
-
-PAGE_PATTERN = "[!_]*.md"
-
-
-@m.parametrize(
-    "test_id, args, expected",
-    [
-        (
-            "no_children",
-            (PAGES["close-waste-transform"],),
-            [],
-        ),
-        (
-            "home",
-            (PAGES["home"],),
-            [PAGES["impeach-vermilion-vacuum"], PAGES["equity-substitute-huddle"]],
-        ),
-        (
-            0,
-            (PAGES["impeach-vermilion-vacuum"],),
-            [
-                PAGES["measure-transient-respite"],
-                PAGES["official-union-advantage"],
-                PAGES["middle-pasture-floating"],
-            ],
-        ),
-        (
-            1,
-            (PAGES["equity-substitute-huddle"],),
-            [
-                PAGES["automatic-party-merit"],
-                PAGES["medium-establish-vital"],
-                PAGES["reaction-diagonal-patter"],
-            ],
-        ),
-        (
-            2,
-            (PAGES["official-union-advantage"],),
-            [
-                PAGES["close-waste-transform"],
-                PAGES["transit-thrum-middle"],
-                PAGES["serpentine-hurry-butcher"],
-            ],
-        ),
-    ],
-)
-def test_get_children(test_id, args, expected):
-    result = wm.get_children(*args)
-    assert result == expected
-
-
-# * ---------------------------------------- * #
-# * get_parent
-
-
-@m.parametrize(
-    "test_id, args, expected",
-    [
-        ("home", (PAGES["home"],), PAGES["home"]),
-        (0, (PAGES["impeach-vermilion-vacuum"],), PAGES["home"]),
-        (1, (PAGES["serpentine-hurry-butcher"],), PAGES["official-union-advantage"]),
-    ],
-)
-def test_get_parent(test_id, args, expected):
-    result = wm.get_parent(*args)
-    assert result == expected
-
-
-# * ---------------------------------------- * #
-# * get_siblings
-
-
-@m.parametrize(
-    "test_id, args, expected",
-    [
-        (
-            "home",
-            (PAGES["home"],),
-            [PAGES["impeach-vermilion-vacuum"], PAGES["equity-substitute-huddle"]],
-        ),
-        (
-            "subpage",
-            (PAGES["impeach-vermilion-vacuum"],),
-            [PAGES["impeach-vermilion-vacuum"], PAGES["equity-substitute-huddle"]],
-        ),
-        ("only_child", (PAGES["slate-slide-course"],), [PAGES["slate-slide-course"]]),
-    ],
-)
-def test_get_siblings(test_id, args, expected):
-    result = wm.get_siblings(*args)
-    assert result == expected
-
-
-# * -------------------------------------------------------------------------------- * #
-# * RELATIVES
-
-
-# * ---------------------------------------- * #
-# * get_prev
-
-
-@m.parametrize(
-    "test_id, args, expected",
-    [
-        (
-            "is_first_child",
-            (
-                PAGES["measure-transient-respite"],
-                [
-                    PAGES["measure-transient-respite"],
-                    PAGES["official-union-advantage"],
-                    PAGES["middle-pasture-floating"],
-                ],
-                0,
-            ),
-            PAGES["impeach-vermilion-vacuum"],
-        ),
-        (
-            "else",
-            (
-                PAGES["official-union-advantage"],
-                [
-                    PAGES["measure-transient-respite"],
-                    PAGES["official-union-advantage"],
-                    PAGES["middle-pasture-floating"],
-                ],
-                1,
-            ),
-            PAGES["measure-transient-respite"],
-        ),
-    ],
-)
-def test_get_prev(test_id, args, expected):
-    result = wm.get_prev(*args)
-    assert result == expected
-
-
-# * ---------------------------------------- * #
-# * get_next_of_last_child
-
-
-@m.parametrize(
-    "test_id, args, expected",
-    [
-        ("is_last_page", (PAGES["reaction-diagonal-patter"],), PAGES["home"]),
-        (
-            "elif_parent_is_last_child",
-            (PAGES["meridian-preserve-winter"],),
-            PAGES["equity-substitute-huddle"],
-        ),
-        (
-            "else",
-            (PAGES["middle-pasture-floating"],),
-            PAGES["equity-substitute-huddle"],
-        ),
-    ],
-)
-def test_get_next_of_last_child(test_id, args, expected):
-    result = wm.get_next_of_last_child(*args)
-    assert result == expected
-
-
-# * ---------------------------------------- * #
-# * get_next
-
-
-@m.parametrize(
-    "test_id, args, expected",
-    [
-        (
-            "has_children",
-            (
-                PAGES["equity-substitute-huddle"],
-                [
-                    PAGES["impeach-vermilion-vacuum"],
-                    PAGES["equity-substitute-huddle"],
-                ],
-                1,
-            ),
-            PAGES["automatic-party-merit"],
-        ),
-        (
-            "elif_is_last_child",
-            (
-                PAGES["serpentine-hurry-butcher"],
-                [
-                    PAGES["close-waste-transform"],
-                    PAGES["transit-thrum-middle"],
-                    PAGES["serpentine-hurry-butcher"],
-                ],
-                2,
-            ),
-            PAGES["middle-pasture-floating"],
-        ),
-        (
-            "else",
-            (
-                PAGES["medium-establish-vital"],
-                [
-                    PAGES["automatic-party-merit"],
-                    PAGES["medium-establish-vital"],
-                    PAGES["reaction-diagonal-patter"],
-                ],
-                1,
-            ),
-            PAGES["reaction-diagonal-patter"],
-        ),
-    ],
-)
-def test_get_next(test_id, args, expected):
-    result = wm.get_next(*args)
-    assert result == expected
-
 
 # * ---------------------------------------- * #
 # * get_nearest
